+++ software/zynq/Synthesizer/SoundComponentsImpl/sine/impl/SineImpl_SW.h
@@ -18,7 +18,7 @@ class SineImpl_SW : public SineSoundComponent {
 
 private:
 
-	float phase;
+	double phase;
 
 public:
